# docs/swagger.yaml
basePath: /api/v2
definitions:
  api.AppInfoResponse:
    properties:
      registration_disabled:
        type: boolean
      socials_disabled:
        type: boolean
      version:
        type: string
      version_sha:
        type: string
    type: object
...
  api.EquipmentResponse:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      default_for:
        items:
          type: string
        type: array
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  api.PaginatedResponse-api_EquipmentResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      page:
        type: integer
      per_page:
        type: integer
      results:
        items:
          $ref: '#/definitions/api.EquipmentResponse'
        type: array
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  api.ProfileSettings:
    properties:
      api_active:
        type: boolean
      api_key:
        type: string
      auto_import_directory:
        type: string
      language:
        type: string
      prefer_full_date:
        type: boolean
      preferred_units:
        $ref: '#/definitions/database.UserPreferredUnits'
      socials_disabled:
        type: boolean
      theme:
        type: string
      timezone:
        type: string
      totals_show:
        type: string
    type: object
...
  api.Response-any:
    properties:
      errors:
        items:
          type: string
        type: array
      results: {}
    type: object
  api.Response-api_AppInfoResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/api.AppInfoResponse'
    type: object
  api.Response-api_EquipmentResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/api.EquipmentResponse'
    type: object
...
  api.Response-api_UserProfileResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/api.UserProfileResponse'
    type: object
...
  api.Response-array_api_UserProfileResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        items:
          $ref: '#/definitions/api.UserProfileResponse'
        type: array
    type: object
...
  api.Response-geojson_FeatureCollection:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/geojson.FeatureCollection'
    type: object
...
  api.UserProfileResponse:
    properties:
      active:
        type: boolean
      admin:
        type: boolean
      birthdate:
        type: string
      created_at:
        type: string
      id:
        type: integer
      language:
        type: string
      last_version:
        type: string
      name:
        type: string
      prefer_full_date:
        type: boolean
      preferred_units:
        $ref: '#/definitions/database.UserPreferredUnits'
      profile:
        $ref: '#/definitions/api.ProfileSettings'
      socials_disabled:
        type: boolean
      theme:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  database.UserPreferredUnits:
    properties:
      distance:
        description: The user's preferred distance unit
        type: string
      elevation:
        description: The user's preferred elevation unit
        type: string
      height:
        description: The user's preferred height unit
        type: string
      speed:
        description: The user's preferred speed unit
        type: string
      weight:
        description: The user's preferred weight unit
        type: string
    type: object
  geojson.Feature:
    properties:
      bbox:
        items:
          type: number
        type: array
      geometry: {}
      id: {}
      properties:
        $ref: '#/definitions/geojson.Properties'
      type:
        type: string
    type: object
  geojson.FeatureCollection:
    properties:
      bbox:
        items:
          type: number
        type: array
      features:
        items:
          $ref: '#/definitions/geojson.Feature'
        type: array
      type:
        type: string
    type: object
  geojson.Properties:
    additionalProperties: true
    type: object
...
info:
  contact: {}
  description: Workout Tracker HTTP API (v2).
  title: Workout Tracker API
  version: "2.0"
paths:
...
  /admin/config:
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_AppInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Update config (admin)
      tags:
      - admin
  /admin/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-array_api_UserProfileResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: List users (admin)
      tags:
      - admin
  /admin/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Delete user (admin)
      tags:
      - admin
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Get user (admin)
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Update user (admin)
      tags:
      - admin
...
  /equipment:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse-api_EquipmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: List equipment
      tags:
      - equipment
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response-api_EquipmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Create equipment
      tags:
      - equipment
  /equipment/{id}:
    delete:
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Delete equipment
      tags:
      - equipment
    get:
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_EquipmentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Get equipment
      tags:
      - equipment
    put:
      consumes:
      - application/json
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_EquipmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Update equipment
      tags:
      - equipment
...
  /workouts/centers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-geojson_FeatureCollection'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Get workout centers
      tags:
      - heatmap
  /workouts/coordinates:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-geojson_FeatureCollection'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Get workout coordinates
      tags:
      - heatmap
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ApiKeyQuery:
    in: query
    name: api-key
    type: apiKey
  CookieAuth:
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"